Update NameUtils.php
There was a reason I removed numbers from being allowable NS characters, but that might have been due to running an ancient version of PHP on Windows as I am unable to reproduce it now.

One issue, however, no identifiers in PHP are allowed to start with numbers (they must be alpha).  I have updated the regex to require the first character of each namespace part be an alpha character.